Create matlab2latex-pretty-toc
Added a Table Of Contents and created rules that enable the definition of multiple header levels. If the MATLAB section header starts with `_` it is seen a subsection, if it starts with `__` it is a subsubsection, in all other cases its a section.
Removed all asterisks from the section commands to enable numbering.